Refactor into two separate files, one for "high-level" and one for "low-level"
We have two interfaces to ExpressionMatrix2 that are under development. The first is a
relatively direct copy of the interface from the C++ EM class. The second is aimed at
being more "pythonic", modifying the interface to meet the expectations of python
users. This change puts them together in a single module where the low-level interface
is optionally available and is a superset of the high-level interface.